More types of colours, hues and saturation
saturation: how much of the color shows through from white
hue: what color the light is

saturation is now a preset with a color, making it possible for pastel colors as well as full colors. furthermore the HTTP request can now send the color values of the light as well as an on/off input.